Create Edge and Contour
MacroNode has Edges. Contour has a doubly-linked list of Edges. There are four subclasses of Contour, which take the advantage of polymoriphism. Three methods place MacroNode on a Contour. Each of these methods has two versions for packing forward and packing backward. MacroNode has verticalDisplacement, which is copied when the MacroNode is copied. MacroNode can be covered by another MacroNode. When deleting a Contour, all Edges whose macroNodes are zero are also deleted to prevent memory leak (delete MacroNode also delete its Edges). There are three legalization methods for Contour, which assume the Contour is used as an interiorContour of ICPTree. Some other methods are designed to be used in placing Macros in ICPTree.