Fix memory leak and discard empty Nodes.
ABSTRACT: Fix the memory leak caused by copying ICPTree and annealing ICPTrees.   DETAILS: 1. In C++, to delete a derived class object by a base class pointer, the destructor of the base class should be VIRTUAL, otherwise the destructor of the derived class will not be called and memory leak occurs. To fix the memory leak, make the destructor of BinaryTree, Node, Contour and State virtual. 2. Empty-Nodes-related code in BinaryTree.copy() and ~BinaryTree() also causes memory leak (but how it happens remains unknown). Remove the code. As a result, BinaryTree now assumes that all Nodes are added by BinaryTree.addNode() and thus adding empty Nodes to an ICPTree will cause crash. 3. Modify ICPTree.initializeMacroNodesOnBranchesRandomly() such that it does not insert empty Nodes and guarantees that there are at least four MacroNodes on branches.