Draw lines between MacroNodes and their children
ABTRACT: Try to fix the issue that Macros are placed out of the bounding box sometimes, but decide to postpone this issue for some future commit.   DETAILS: Draw lines connecting MacroNodes by adding Window.drawLine() and modify TraversalTaskDrawMacroNode.doWhenTraversing() to show which part of the ICPTree are uncorrectly placed.   OBSERVATIONS: The crazily placed Macros seems to come from the right subtree of the root.