Add MacroNode.spacing
ABSTRACT: Add spacing between on-branch MacroNodes to replace empty Nodes.   DETAILS: 1. Currently, always place the spacing to the right (in terms of TopContour) of the Macros. 2. In MacroNode, add spacing, spacingDirection and usingSpacing. Add setMacroXStart() and other three to consider the spacing when placing the Macro. Modify copy() and createTopEdge() and other three. 3. In Contour, add setMacroNodeMacroXStart() and other three to use MacroNode.setMacroXStart(). Call these methods in replace of setMacroXStart() when placing Macros on tails or heads. 4. In ICPTree, add setAllMacroNodesNotUsingSpacing() and setMacroNodesOnBranchesSpacingDirection(), then use the spacing of Corner 1, 2, 3 in placingMacros. Add changeRangeOfSpacing. Modify calculateInteriorRegionArea() to replace empty Nodes with spacing.   SEE the Trello card, Add MacroNode.spacing, for more details.