Modify bounding box, interior region area and add desired area
ABSTRACT: Able to set the target area of bounding box area. Let interior region area consider the spacing of MacroNodes, and remove empty-Nodes-related code, and consider the interiorRightContour. Able to set the desired area of Floorplan.    DETAILS: 1.1. Add a parameter, targetArea, to BoundingBoxArea(). 1.2. Add Floorplan.calculateMacrosArea() to calculate the total area of preplaced and movable Macros. 1.3. Fix a big bug: change the return type of ICPTree.ICPTree.getBoundingBoxAspectRatio() from int to double. 2.1. Modify ICPTree.calculateInteriorRegionArea() to replace empty Nodes with MacroNode.spacing. 2.2. Use interiorRightContour.head as the left boundary of the interior region. 3. Add Floorplan.setDesiredRegion() and make it as a cost in sa such that the Macros tend to be placed inside the desired region (however the result is not as good as expected).   SEE Trello cards, Modify BoundingBoxArea to subtract minimum area, Modify the calculation of interior region area, and Set desired region of floorplan.