Initialize a balanced ICPTree in the desired region
ABTRACT: Initialize a balanced ICPTree, which places the Macros around the rectangular interior region and distributes the MacroNodes evenly in the four subtree according to the subtree area within the desired region.   ALGORITHM: 1. Calculate the interior region position. 2. Select and assign BranchNodes and CornerNodes and place them around the interior region. 3. Assign the others as NormalNodes and insert them to the four subtrees such that the Macros area of each subtree matches the max area. Whether a MacroNode is inserted as leftNode or rightNode is determined randomly according to the aspect ratio of the Macro.   DETAILS SEE Trello card, Initialize a balanced ICPTree in the desired region.